Implement strict or loose parameter bounds
Loose bounds can be helpful if participants show opposite framing effects (e.g. more LL choices on delay than date frames) or when best-fitting parameter values are at bounds. The model can than be fit with the
following bounds in order to see whether this improves the fit to the data.